feat: implement OpenMailApp functionality with new email composition and app selection features
