+++ lib/src/constants.dart
@@ -1,0 +1,11 @@
+/// Launch Schemes for supported apps:
+const String kLaunchSchemeAppleMail = 'message://';
+const String kLaunchSchemeGmail = 'googlegmail://';
+const String kLaunchSchemeDispatch = 'x-dispatch://';
+const String kLaunchSchemeSpark = 'readdle-spark://';
+const String kLaunchSchemeAirmail = 'airmail://';
+const String kLaunchSchemeOutlook = 'ms-outlook://';
+const String kLaunchSchemeYahoo = 'ymail://';
+const String kLaunchSchemeFastmail = 'fastmail://';
+const String kLaunchSchemeSuperhuman = 'superhuman://';
+const String kLaunchSchemeProtonmail = 'protonmail://';
