+++ lib/src/models.dart
@@ -1,0 +1,7 @@
+// Reeksportuje wszystkie klasy z katalogu src
+export 'compose_data.dart';
+export 'constants.dart';
+export 'email_content.dart';
+export 'mail_app.dart';
+export 'mail_app_picker_dialog.dart';
+export 'open_mail_app_result.dart';
